consolidated codependent models into single file
